[연습문제5] 작성일 : 2013.7.29_14:00 작성자 : 이영남 변경내용 	현재 존재하는 말 정렬할때 while문 버그 수정

+++ src/pieces/Piece.java
@@ -250,8 +250,8 @@ public class Piece  {
 
 
 
-	public int compare(Piece target) {
-		return (int) (this.getPoint() - target.getPoint());
+	public double compare(Piece target) {
+		return (this.getPoint() - target.getPoint());
 	}
 	
 }
